Added functions menu and started search tweets
I didn't use the same type of menu method as before, I think this way is a little simpler.
Functions will continue to run as long as user doesn't logout. 
Also made it so when registering, if register was successful they then need to login.
Search tweet is still a work in progress, didn't have time to finish and test it.